setup eslint, prettier, and travis ci
