+++ package.json
@@ -4,7 +4,8 @@
   "description": "Uvic Web Scrapper",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "lint": "./node_modules/.bin/eslint src/**/*.js"
   },
   "author": "Joel Kerfoot",
   "license": "ISC",
@@ -13,5 +14,9 @@
     "form-data": "^2.5.1",
     "request": "^2.88.0",
     "request-promise": "^4.2.4"
+  },
+  "devDependencies": {
+    "eslint": "^6.6.0",
+    "prettier": "^1.18.2"
   }
 }
